BusinessManager: Add CRUD interfaces
A CRUD interface is added for the models. It has been implemented
making use of templates.

A CRUD interface is added for the views.

Also a new type has been created for list all the CRUD options. It
contains an extra option which must be deleted in the final
release.