BusinessManager: Implement Mainwindow and Bills
Mainwindow and Bills are added.

Fix some arrows in the architecture diagram.
Unit test for Bills created. Added a fixme to investigate in the
future how to delete the singleton for each test case to make the
test cases independent.

Updated how to manage the program compilation. A new directory for
config files is created. It will contain everything which is need
to compile each project.

Include directories are modified due to they are added in the
makefile of each project.